# docs/swagger.yaml
basePath: /api
definitions:
  apperror.AppError:
    properties:
      message:
        type: string
    type: object
  auth.AuthResponse:
    properties:
      accessToken:
        type: string
      user:
        $ref: '#/definitions/user.User'
    type: object
  auth.CodeRequest:
    properties:
      code:
        type: string
      email:
        type: string
    required:
    - code
    - email
    type: object
  auth.EmailPasswordRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  auth.EmailRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  auth.JwtToken:
    properties:
      accessToken:
        type: string
    type: object
  auth.PasswordRequest:
    properties:
      password:
        minLength: 8
        type: string
    required:
    - password
    type: object
  auth.ProfileRequest:
    properties:
      firstName:
        maxLength: 50
        minLength: 1
        type: string
      lastName:
        maxLength: 50
        minLength: 1
        type: string
      username:
        maxLength: 30
        minLength: 3
        type: string
    required:
    - firstName
    - lastName
    - username
    type: object
...
  country.Country:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  handler.ContriesResponse:
    properties:
      countries:
        items:
          $ref: '#/definitions/country.Country'
        type: array
    type: object
...
  handler.RegionsResponse:
    properties:
      regions:
        items:
          $ref: '#/definitions/region.Region'
        type: array
    type: object
  handler.StatesResponse:
    properties:
      states:
        items:
          $ref: '#/definitions/state.State'
        type: array
    type: object
...
  region.Region:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  state.State:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  user.User:
    properties:
      age:
        type: integer
      avatar:
        type: string
      country:
        $ref: '#/definitions/country.Country'
      email:
        type: string
      firstName:
        type: string
      hasBusinessProfile:
        type: boolean
      id:
        type: integer
      isAdmin:
        type: boolean
      isPasswordSet:
        type: boolean
      isVerified:
        type: boolean
      lastName:
        type: string
      phoneNumber:
        type: string
      region:
        $ref: '#/definitions/region.Region'
      state:
        $ref: '#/definitions/state.State'
      username:
        type: string
    type: object
  user.UserResponse:
    properties:
      user:
        $ref: '#/definitions/user.User'
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Server for Kushfinds application
  title: Kushfinds API
  version: "1.0"
paths:
  /auth/login/email:
    post:
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.EmailRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.AppError'
      tags:
      - auth
  /auth/login/password:
    post:
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.EmailPasswordRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.AppError'
      tags:
      - auth
...
  /auth/refresh:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.JwtToken'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.AppError'
      tags:
      - auth
  /auth/register/email:
    post:
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.EmailRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.AppError'
      tags:
      - auth
  /auth/register/password:
    patch:
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.PasswordRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - ApiKeyAuth: []
      tags:
      - auth
  /auth/register/profile:
    patch:
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.ProfileRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - ApiKeyAuth: []
      tags:
      - auth
  /auth/register/verify:
    post:
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.CodeRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.AppError'
      tags:
      - auth
  /auth/verify/resend:
    post:
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.EmailRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.AppError'
      tags:
      - auth
  /countries:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ContriesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.AppError'
      tags:
      - location
  /countries/{country_id}/states:
    get:
      parameters:
      - description: Country ID
        in: path
        name: country_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.StatesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.AppError'
      tags:
      - location
...
  /states/{state_id}/regions:
    get:
      parameters:
      - description: State ID
        in: path
        name: state_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RegionsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.AppError'
      tags:
      - location
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"